updating dayidsub to ts
